+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- `option().groupSwitch()`, which works like `groupChoice()`, but uses a `switch()` option rather than a `choice()` option.
+- `UsageError` now has a `statusCode` parameter (which defaults to 1). If you're using `ClicktCommand.main`, the value of `statusCode` will be passed to `exitProcess`. 
+
 ### Changed
 - Shell completion code is now printed by throwing a `PrintCompletionMessage` (a subclass of `PrintMessage`) rather than calling `echo` directly.
-
-### Added
-- `UsageError` now has a `statusCode` parameter (which defaults to 1). If you're using `ClicktCommand.main`, the value of `statusCode` will be passed to `exitProcess`. 
 
 ## [2.2.0] - 2019-09-25
 ### Added